+++ src/common/warn.ts
@@ -1,0 +1,5 @@
+export function warn(message: string) {
+    if (process.env.NODE_ENV !== "production") {
+        console.warn(`[react-router-typesafe-routes] ${message}`);
+    }
+}